requirements.txt
python-telegram-bot==20.7
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
aiohttp==3.9.1
Pillow==10.1.0
pytesseract==0.3.10
regex==2023.12.25
